version final de mi base de datos en mongodb

+++ backend/controllers/tareasControllers.js
@@ -3,7 +3,7 @@ const Tarea =require('../models/tareasModel')
 
 
 const getTareas =asyncHandler(async (req, res) => {
-    const tareas = await Tarea.find()
+    const tareas = await Tarea.find({user: req.user.id})
     res.status(200).json(tareas)
 })
 
@@ -14,7 +14,8 @@ const createTareas = asyncHandler( async (req, res) => {
     }
     
     const tarea = await Tarea.create({
-        texto : req.body.texto
+        texto : req.body.texto,
+        user:req.user.id
     }
 
     )
@@ -24,30 +25,44 @@ const createTareas = asyncHandler( async (req, res) => {
 
 const updateTareas = asyncHandler( async (req, res) => {
     const tarea = await Tarea.findById(req.params.id)
+
     if(!tarea){
         res.status(400)
         throw new Error ('Tarea no encontrada')
     }
+    
+    if(tarea.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error ('Usuario no autorizado')
+    }else{
+        const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id , req.body, {new: true})
 
-    const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id , req.body, {new: true})
-
-    res.status(200).json(tareaUpdated)
-
-
+        res.status(200).json(tareaUpdated)
+    }
 })
 
 
 const deleteTareas = asyncHandler( async (req, res) => {
 
+    //verificamos que la tarea existe
     const tarea = await Tarea.findById(req.params.id)
+
     if(!tarea){
         res.status(400)
         throw new Error ('Tarea no encontrada')
     }
+//verificamos que la tarea pertenece al usuario 
+    if(tarea.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error ('Usuario no autorizado')
+    }else{
+        await tarea.deleteOne()
 
-    await tarea.deleteOne()
+        res.status(200).json({id: req.params.id})
+    }
 
-    res.status(200).json({id: req.params.id})
+
+
 })
